+++ onnx/version_converter/adapters/broadcast_forward_compatibility.h
@@ -35,7 +35,7 @@ class BroadcastForwardCompatibility final : public Adapter {
             if (target_version().version() >= 13){ //Unsqueeze takes 'axes' input
               Tensor t;
               t.elem_type() = TensorProto_DataType_INT64;
-              t.sizes() = std::vector<int64_t>{axes.size()};
+              t.sizes() = std::vector<int64_t>{static_cast<int64_t>(axes.size())};
               auto& data = t.int64s();
               for (auto a : axes) {
                 data.emplace_back(a);
