+++ onnx/version_converter/adapters/split_12_13.h
@@ -17,7 +17,7 @@ class Split_12_13 : public Adapter {
       std::vector<int64_t> axes) const {
     Tensor t;
     t.elem_type() = TensorProto_DataType_INT64;
-    t.sizes() = std::vector<int64_t>{axes.size()};
+    t.sizes() = std::vector<int64_t>{static_cast<int64_t>(axes.size())};
     auto& data = t.int64s();
     for (auto a : axes) {
       data.emplace_back(a);
